Refactor: SearchRepositoryのディレクトリを変更
